feat(message-actions): move to separate components
